+++ routes/loans.py
@@ -11,30 +11,21 @@ def get_loans():
     return jsonify(ret)
 
 
-@app.route("/loans/<loan_id>", methods=["GET"])
+@app.route("/loans/<loan_id>", methods=["GET", "PUT", "DELETE"])
 def get_loan(loan_id):
-    loan = Loan.query.get_or_404(loan_id)
-    ret = loan.to_json()
-    return jsonify(ret)
-
-
-@app.route("/loans/<loan_id>", methods=["PUT"])
-def update_loan(loan_id):
-    loan = Loan.query.get_or_404(loan_id)
-    user = User.query.get_or_404(request.form["user"])
-    book = Book.query.get_or_404(request.form["book"])
-    loan.due_date = request.form["dueDate"]
-    loan.user = user
-    loan.book = book
-    db.session.add(loan)
-    db.session.commit()
-    ret = loan.to_json()
-    return jsonify(ret)
-
-
-@app.route("/loans/<loan_id>", methods=["DELETE"])
-def delete_loan(loan_id):
-    loan = Loan.query.get_or_404(loan_id)
-    db.session.delete(loan)
-    db.session.commit()
-    return jsonify("")
+    if request.method == "GET":
+        loan = Loan.query.get_or_404(loan_id)
+        ret = loan.to_json()
+        return jsonify(ret)
+    elif request.method == "PUT":
+        loan = Loan.query.get_or_404(loan_id)
+        loan.due_date = datetime.fromisoformat(request.json["dueDate"])
+        db.session.add(loan)
+        db.session.commit()
+        ret = loan.to_json()
+        return jsonify(ret)
+    elif request.method == "DELETE":
+        loan = Loan.query.get_or_404(loan_id)
+        db.session.delete(loan)
+        db.session.commit()
+        return jsonify("")
